Разбила код на отдельные модули
